1.05

+++ projet_reco.py
@@ -11,10 +11,14 @@ from pylti.flask import lti
 import MySQLdb
 from get_params import get_params
 import xlrd
+from get_exo import get_exo
 
-VERSION = '0.0.5'
+reload(sys)
+sys.setdefaultencoding('utf8')
+VERSION = '1.03'
 UPLOAD_FOLDER='exos/'
 ALLOWED_EXTENSIONS=set(['xlsx'])
+ALLOWED_EXTENSIONS2=set(['tex'])
 app = Flask(__name__)
 app.config['UPLOAD_FOLDER']=UPLOAD_FOLDER
 app.config.from_object('config')
@@ -74,6 +78,10 @@ def index_staff(lti=lti):
 def allowed_file(filename):
     return '.' in filename and \
            filename.rsplit('.', 1)[1].lower() in ALLOWED_EXTENSIONS
+
+def allowed_file2(filename):
+    return '.' in filename and \
+           filename.rsplit('.', 1)[1].lower() in ALLOWED_EXTENSIONS2
     
 @app.route('/upload_exo', methods=['GET','POST'])
 @lti(request='session', error=error, role='staff', app=app)
@@ -103,42 +111,55 @@ def upload_exo(lti=lti):
 def majDB(lti=lti):
     database = MySQLdb.connect(host="127.0.0.1",port=3306,user="root",passwd="",db="moodle")
     cHandler = database.cursor() 
-    cHandler.execute("CREATE TABLE IF NOT EXISTS mdl_exos_recommendation (num_exo integer, theme integer, savoir_faire integer, cours_id integer)")
-    cHandler.execute("DELETE FROM mdl_exos_recommendation")    
+    cHandler.execute("CREATE TABLE IF NOT EXISTS mdl_exos_recommendation (num_exo integer, id_theme integer, id_savoir_faire integer, cours_id integer)")
+    cHandler.execute("DELETE FROM mdl_exos_recommendation WHERE cours_id=%s",lti.user_id)
+    cHandler.execute("CREATE TABLE IF NOT EXISTS mdl_comp_recommendation (id_savoir_faire integer, savoir_faire text, cours_id integer) ")  
+    cHandler.execute("DELETE FROM mdl_comp_recommendation WHERE cours_id=%s",lti.user_id)
+    cHandler.execute("CREATE TABLE IF NOT EXISTS mdl_theme_recommendation (id_theme integer, theme text, cours_id integer) ")  
+    cHandler.execute("DELETE FROM mdl_theme_recommendation WHERE cours_id=%s",lti.user_id)
     # Open the workbook and define the worksheet
     book = xlrd.open_workbook("./exos/bdd.xlsx")
-    sheets=book.sheet_names()
+    sheets = book.sheet_names()
+    sheet0 = book.sheet_by_name(sheets[0])
     sheet1 = book.sheet_by_name(sheets[1])
     sheet2 = book.sheet_by_name(sheets[2])
+    sheet3 = book.sheet_by_name(sheets[3])
     # Get the cursor, which is used to traverse the database, line by line
     cursor = database.cursor()
     cours_id = lti.user_id
-    # Create the INSERT INTO sql query
-    query = """INSERT INTO mdl_exos_recommendation (num_exo, theme, cours_id) VALUES (%s, %s, %s)"""
-#    WHERE NOT EXISTS (SELECT * FROM mdl_exos_recommendation WHERE num_exo=%s AND theme=%s)"""
-    
     # Tableau pour stocker les différents savoirs faire associés aux exercices
     tab_sf=[]
     tab_sf.append([])
-    num_exo=[]
+    tab_sft=[]
+    tab_sft.append([])
+    tab_t=[]
+    tab_t.append([])
     theme=[]
-    # Create a For loop to iterate through each row in the XLS file, starting at row 2 to skip the headers
-#    for r in range(1, sheet1.nrows):
-#        num_exo = int(sheet1.cell(r,0).value)
-#        
-#        cursor.execute(query,(num_exo,theme,cours_id))
+    #Trie les différents savoirs et les attache aux exercices concernés
     for r in range(1, sheet2.nrows):
         tab_sf.append([])
         tab_sf[int(sheet2.cell(r,0).value)].append(int(sheet2.cell(r,1).value))
-    
-    #Remplissage des savoir_faire
+    #Trie les savoir_faire et les attache à leur id
+    for r in range(1,sheet3.nrows):
+        tab_sft.append([])
+        tab_sft[int(sheet3.cell(r,0).value)].append(sheet3.cell(r,2).value)
+    #Associe les themes a leur id
+    for r in range(1,sheet0.nrows):
+        tab_t.append([])
+        tab_t[int(sheet0.cell(r,1).value)].append(sheet0.cell(r,0).value)
+    #Remplissage de la db
     for r in range(1,sheet1.nrows):
         theme = int(sheet1.cell(r,1).value)
         for j in range(0,len(tab_sf[r])):
             value=tab_sf[r][j]
-            cursor.execute("INSERT INTO mdl_exos_recommendation (num_exo, savoir_faire, cours_id, theme) VALUES (%s,%s,%s,%s)", \
-                       (r,value,cours_id,theme))      
-        
+            cursor.execute("INSERT INTO mdl_exos_recommendation (num_exo, id_savoir_faire, cours_id, id_theme) VALUES (%s,%s,%s,%s)", \
+                       (r,value,cours_id,theme)) 
+    for r in range(1,sheet3.nrows):
+        for j in range(0,len(tab_sft[r])):
+            cursor.execute("INSERT INTO mdl_comp_recommendation (id_savoir_faire, savoir_faire, cours_id) VALUES (%s,%s,%s)",(r,tab_sft[r][j],lti.user_id))
+    for r in range(1,sheet0.nrows):
+        for j in range(0,len(tab_t[r])):
+            cursor.execute("INSERT INTO mdl_theme_recommendation (id_theme, theme, cours_id) VALUES (%s,%s,%s)",(r,tab_t[r][j],lti.user_id))
     # On supprime le fichier telechargé pour ne pas avoir de conflits lors d'une mise a jour
     #os.remove(".exos/bdd.xlsx")
     
@@ -151,6 +172,31 @@ def majDB(lti=lti):
     # Close the database connection
     database.close()
     return render_template("upload_reussit.html",lti=lti)
+
+@app.route('/up_latex', methods=['GET','POST'])
+@lti(request='session', error=error, role='staff', app=app)
+def upload_latex(lti=lti):
+    # L'utilisateur entre son fichier latex
+    return render_template('up_latex.html',lti=lti)
+    
+@app.route('/latex', methods=['GET','POST'])
+@lti(request='session', error=error, role='staff', app=app)
+def latex(lti=lti):
+    data=request.files['file']
+    if data == '':
+        flash('Entrez un fichier latex (.tex)')
+        return redirect(request.url)
+    if data and allowed_file2(data.filename):
+        data.save(os.path.join(app.config['UPLOAD_FOLDER'], 'exos.tex'))
+#        data=open('.\exos\exos.tex').read()
+#        database = MySQLdb.connect(host="127.0.0.1",port=3306,user="root",passwd="",db="moodle")
+#        cHandler = database.cursor() 
+#        cours_id = lti.user_id
+#        cHandler.execute("CREATE TABLE IF NOT EXISTS mdl_exos_recommendation_latex (fichier_latex blob, cours_id integer)")
+#        cHandler.execute("DELETE FROM mdl_exos_recommendation_latex WHERE cours_id=%s",lti.user_id)
+#        cHandler.execute("INSERT INTO mdl_exos_recommendation_latex (fichier_latex, cours_id) VALUES (%s,%s)",(data,cours_id))
+        return render_template('ok_latex.html',lti=lti)
+    return render_template('up_latex.html',lti=lti)
     
 @app.route('/see_competences', methods=['GET','POST'])
 @lti(request='session', error=error, role='staff', app=app)
@@ -170,11 +216,13 @@ def competences(lti=lti):
     database = MySQLdb.connect(host="127.0.0.1",port=3306,user="root",passwd="",db="moodle")
     cHandler = database.cursor() 
     cours_id = lti.user_id
-    cHandler.execute("SELECT DISTINCT savoir_faire FROM mdl_exos_recommendation WHERE num_exo=%s AND cours_id=%s",(data,cours_id))
+    cHandler.execute("SELECT DISTINCT id_savoir_faire FROM mdl_exos_recommendation WHERE num_exo=%s AND cours_id=%s",(data,cours_id))
     results=cHandler.fetchall()
     resultats=[]
     for items in results[1:]:
-        resultats.append(str(items[0]))
+        cHandler.execute("SELECT savoir_faire FROM mdl_comp_recommendation WHERE id_savoir_faire=%s AND cours_id=%s",(items[0],cours_id))
+        viv=cHandler.fetchall()
+        resultats.append(viv[0][0].decode("latin1"))
     return render_template("competences.html",lti=lti, resultats=resultats,num=data)
     
 @app.route('/see_exos', methods=['GET','POST'])
@@ -201,7 +249,7 @@ def exos(lti=lti):
         results=cHandler.fetchall()
         resultats=[]
         for items in results:
-            resultats.append(str(items[0]))
+            resultats.append(get_exo(str(items[0])))
         return render_template("exos.html",lti=lti, resultats=resultats,num1=data1)
     if data1 and not data2=='':
         cHandler.execute("SELECT m1.num_exo FROM mdl_exos_recommendation m1 JOIN\
@@ -211,7 +259,7 @@ def exos(lti=lti):
         results=cHandler.fetchall()
         resultats=[]
         for items in results:
-            resultats.append(str(items[0]))
+            resultats.append(get_exo(str(items[0])))
         return render_template("exos2.html",lti=lti, resultats=resultats,num1=data1, num2=data2)
     return render_template('see_exos.html',lti=lti)
